requirements.txt
asgiref==3.4.1
Django==3.2.13
django-environ==0.8.1
django-jazzmin==2.5.0
gunicorn==20.1.0
Pillow==8.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
